+++ tools/soundOptions/tool.js
@@ -6,7 +6,7 @@
         name: 'soundOptions',
         category: 'general',
         type: 'multiple',
-        possibleValues: ['noMove'],
+        possibleValues: ['noMove', 'flickerSound'],
         defaultValue: false,
         advanced: true
       },
@@ -41,6 +41,7 @@
         'options.soundVolume': 'Sound volume (0-100)',
         'options.soundVoice': 'Speech voice',
         'soundOptions.noMove': 'No move sounds',
+        'soundOptions.flickerSound': 'Time flicker sound',
         'options.timeAlert': 'Time alert (minutes)',
         'timeAlert.s30': '0:30',
         'timeAlert.s60': '1:00',
@@ -57,6 +58,7 @@
         'options.soundVolume': 'Volum sonor (0-100)',
         'options.soundVoice': 'Voce folosit\u0103',
         'soundOptions.noMove': 'F\u0103r\u0103 sunet la mutare',
+        'soundOptions.flickerSound': 'Sunet la salt de timp',
         'options.timeAlert': 'Alert\u0103 timp (minute)',
         'timeAlert.s30': '0:30',
         'timeAlert.s60': '1:00',
@@ -82,6 +84,14 @@
       }
     };
 
+    makeFlickerSound = (seconds) => {
+      const lt = this.lichessTools;
+      const $ = lt.$;
+      if (this.options.flickerSound) {
+        lt.play('other/failure2.mp3');
+      }
+    };
+
     checkClock = ()=>{
       const lt = this.lichessTools;
       const $ = lt.$;
@@ -90,6 +100,13 @@
       const m = /^\s*(?:(?<h>\d+):)?(?<m>\d+):(?<s>\d+(?:\.\d+)?)\s*$/.exec(timeStr);
       if (!m) return;
       const time = (+m.groups.h||0)*3600 + (+m.groups.m||0)*60 + (+m.groups.s||0);
+      if (this.options.flickerSound) {
+        if (time<this.prevTime-1) {
+          this.makeFlickerSound();
+        }
+      }
+      this.prevTime = time;
+
       for (let i=this.options.times.length - 1; i>=0; i--) {
         const t = this.options.times[i];
         if (!t.enabled) continue;
@@ -147,6 +164,7 @@
       this.logOption('Time alert', timeAlert);
       this.options = {
         noMove: lt.isOptionSet(soundOptions, 'noMove'),
+        flickerSound: lt.isOptionSet(soundOptions, 'flickerSound'),
         times: [30,60,90,120,180,300].map(s=>({
           seconds: s,
           enabled: lt.isOptionSet(timeAlert, 's'+s)
@@ -168,7 +186,7 @@
         .observer()
         .off('.rclock-bottom *',this.checkClock);
       if ($('.playing .round__app').length) {
-        const hasTimeAlert = this.options.times.find(t=>t.enabled);
+        const hasTimeAlert = this.options.flickerSound || this.options.times.find(t=>t.enabled);
         if (hasTimeAlert) {
           $('.round__app')
             .observer()